Step1 - 구간 추가 요구사항 반영 (#838)
* chore: 1주 차 코드 가져오기

* test: Line.addSection() 테스트 작성

* refactor: 구간 추가 유닛 테스트 리팩터링

* test: 구간 조회, 구간 제거 유닛 테스트 작성

* refactor: 지하철 삭제 리팩터링

* refactor: 지하철 역의 id 값으로 검증하던 것을 Sections 동등성 비교로 수정

* test: LineServiceTest 작성

* test: LineServiceMockTest 작성

* refactor: Line 유닛 테스트 LineTestFixture 클래스 분리

* docs: 기능 요구사항 정리

* test: 기존 구간 추가 인수테스트 수정 및 필요없어진 테스트 케이스 제거

* feat: 노선 구간 노선 가운데 추가 기능

* refactor: 사용하지 않는 예외 제거

* refactor: 가운데 구간 추가 로직 가독성을 위해 메서드 분리

* feat: 구간 가운데 구간 추가 시 길이 검사 추가

* chore: 실수로 빠진 fetch = FetchType.LAZY 추가

* feat: 구간 처음에 구간을 추가하는 기능

* refactor: 구간 추가 메서드 추출

* refactor: 구간 추가 테스트 간소화

* feat: Sections 생성 시 첫 구간 판단을 Sections 본인이 하도록 수정

* refactor: 구간 추가 = addSection 으로 통일

* feat: 이미 등록된 역은 노선 구간에 등록될 수 없도록 수정

* refactor: 불필요한 코드 제거

* feat: 구간 추가시 중복 에러가 발생하면 400 코드를 응답

* refactor: EOF 추가 및 사용하지 않는 코드 제거

* fix: 첫 구간과 마지막 구간을 찾는 방식 오류 수정

+++ src/test/java/nextstep/subway/utils/DatabaseCleanup.java
@@ -1,5 +1,6 @@
 package nextstep.subway.utils;
 
+import nextstep.subway.acceptance.line.LineAcceptanceTest;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
@@ -11,7 +12,7 @@ import javax.persistence.PersistenceContext;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Profile("test")
+@Profile("databaseCleanup")
 @Service
 public class DatabaseCleanup implements InitializingBean {
     @PersistenceContext
@@ -28,13 +29,20 @@ public class DatabaseCleanup implements InitializingBean {
     }
 
     @Transactional
-    public void execute() {
+    public void execute(Object testClass) {
         entityManager.flush();
         entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY FALSE").executeUpdate();
         for (String tableName : tableNames) {
-            entityManager.createNativeQuery("TRUNCATE TABLE " + tableName).executeUpdate();
-            entityManager.createNativeQuery("ALTER TABLE " + tableName + " ALTER COLUMN ID RESTART WITH 1").executeUpdate();
+            truncateTables(testClass, tableName);
         }
         entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY TRUE").executeUpdate();
+    }
+
+    private void truncateTables(Object testClass, String tableName) {
+        if (testClass.equals(LineAcceptanceTest.class) && tableName.equals("Station")) {
+            return;
+        }
+        entityManager.createNativeQuery("TRUNCATE TABLE " + tableName).executeUpdate();
+        entityManager.createNativeQuery("ALTER TABLE " + tableName + " ALTER COLUMN ID RESTART WITH 1").executeUpdate();
     }
 }
